Moved backend.py contents to __init__.py
